support getParameters in other handlers
